renamed aspect packages (in prep to migration to seperate npm packages)
