+++ src/util/feature-u/aspect/feature-redux/index.js
@@ -2,14 +2,14 @@ import reducerAspect  from './reducerAspect';
 import slicedReducer  from './slicedReducer';
 
 //*** 
-//*** Promote all feature-u-redux utilities through this centralized module.
+//*** Promote all feature-redux utilities through this centralized module.
 //*** 
 
 // NOTE: This non-default export supports ES6 imports.
 //       Example:
-//         import { reducerAspect }    from 'feature-u-redux';
+//         import { reducerAspect }    from 'feature-redux';
 //       -or-
-//         import * as FeatureURedux from 'feature-u-redux';
+//         import * as FeatureURedux from 'feature-redux';
 export {
   reducerAspect,
   slicedReducer,
@@ -17,9 +17,9 @@ export {
 
 // NOTE: This default export supports CommonJS modules (otherwise Babel does NOT promote them).
 //       Example:
-//         const { reducerAspect } = require('feature-u-redux');
+//         const { reducerAspect } = require('feature-redux');
 //       -or-
-//         const FeatureURedux = require('feature-u-redux');
+//         const FeatureURedux = require('feature-redux');
 export default {
   reducerAspect,
   slicedReducer,
