+++ src/util/feature-u/docs/api.md
@@ -1,4 +1,4 @@
-# feature-u API
+feature-u API
 <a name="createFeature"></a>
 
 ## createFeature(name, [enabled], [publicFace], [appWillStart], [appDidStart], [pluggableAspects]) ⇒ Feature
@@ -28,7 +28,7 @@ Add additional Feature keyword (typically used by Aspect extensionsto Feature).
 <a name="managedExpansion"></a>
 
 ## managedExpansion(managedExpansionCB) ⇒ [`managedExpansionCB`](#managedExpansionCB)
-Mark the supplied managedExpansionCB as a "managed expansioncallback", distinguishing it from other functions (such as reducerfunctions).Features may communicate AspectContent directly, or through amanagedExpansionCB.  The latter: 1. supports cross-feature communication (through app object    injection), and  2. minimizes circular dependency issues (of ES6 modules).Managed Expansion Callbacks are used when a fully resolved `app`object is requried during in-line code expansion.  They are merelyfunctions that are passed the `app` object and return theexpanded AspectContent (ex: reducer, logic modules, etc.).The managedExpansionCB function should conform to the followingsignature:```jsAPI: managedExpansionCB(app): AspectContent```Example (feature-u-redux `reducerAspect`):```js  export default slicedReducer('foo', managedExpansion( (app) => combineReducers({...reducer-code-requiring-app...} ) ));```SideBar: For reducer aspects, slicedReducer() should always wrap         the the outer function passed to createFunction(), even         when managedExpansion() is used.Example (feature-u-redux-logic `logicAspect`):```js  export const startAppAuthProcess = managedExpansion( (app) => createLogic({    ...logic-code-requiring-app...  }));```Please refer to the feature-u `managedExpansion()` documentationfor more detail.
+Mark the supplied managedExpansionCB as a "managed expansioncallback", distinguishing it from other functions (such as reducerfunctions).Features may communicate AspectContent directly, or through amanagedExpansionCB.  The latter: 1. supports cross-feature communication (through app object    injection), and  2. minimizes circular dependency issues (of ES6 modules).Managed Expansion Callbacks are used when a fully resolved `app`object is requried during in-line code expansion.  They are merelyfunctions that are passed the `app` object and return theexpanded AspectContent (ex: reducer, logic modules, etc.).The managedExpansionCB function should conform to the followingsignature:```jsAPI: managedExpansionCB(app): AspectContent```Example (feature-redux `reducerAspect`):```js  export default slicedReducer('foo', managedExpansion( (app) => combineReducers({...reducer-code-requiring-app...} ) ));```SideBar: For reducer aspects, slicedReducer() should always wrap         the the outer function passed to createFunction(), even         when managedExpansion() is used.Example (feature-redux-logic `logicAspect`):```js  export const startAppAuthProcess = managedExpansion( (app) => createLogic({    ...logic-code-requiring-app...  }));```Please refer to the feature-u `managedExpansion()` documentationfor more detail.
 
 
 | Param | Type | Description |
